+++ src/RequestData.php
@@ -2,6 +2,8 @@
 
 namespace React\HttpClient;
 
+use InvalidArgumentException;
+
 class RequestData
 {
     private $method;
@@ -30,6 +32,16 @@ class RequestData
             $connectionHeaders,
             $headers
         );
+    }
+
+    public function getMethod()
+    {
+        return strtoupper($this->method);
+    }
+
+    public function getUrl()
+    {
+        return $this->url;
     }
 
     public function getScheme()
@@ -77,5 +89,35 @@ class RequestData
         $data .= "\r\n";
 
         return $data;
+    }
+
+    /**
+     * Processes a redirect request, updating internal values to represent the new request data.
+     *
+     * @param integer $code     HTTP status code for the redirect.
+     * @param string  $location The Location header received.
+     */
+    public function redirect($code, $location)
+    {
+        switch ($code) {
+
+            //These cases require that we switch to the GET method.
+            //@see https://github.com/bagder/curl/blob/cc28bc472ec421cec2ba26d653e53892998a248d/lib/transfer.c#L1736
+            case 301:
+            case 302:
+                $this->method = 'GET';
+
+            //Note: 303, 307, 308 status is not supported in HTTP/1.0.
+            // case 303:
+            // case 307:
+            // case 308:
+
+                //Of course switch the location.
+                $this->url = $location;
+                break;
+
+            default:
+                throw new InvalidArgumentException(sprintf("Redirect code %u is not supported", $code));
+        }
     }
 }